backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
PyPDF2==3.0.1
google-generativeai>=0.4.0
python-dotenv>=1.0.0
pymongo>=4.6.0
dnspython>=2.5.0
openai>=1.0.0
